Refactor: factor out conveyor.rs
